+++ deps/MongoBson/bson/util/builder.h
@@ -188,8 +188,8 @@ namespace mongo {
         }
 
         void appendStr(const StringData &str , bool includeEndingNull = true ) {
-            const int len = str.size() + ( includeEndingNull ? 1 : 0 );
-            memcpy(grow(len), str.data(), len);
+            const size_t len = str.size() + ( includeEndingNull ? 1 : 0 );
+            memcpy(grow((int)len), str.data(), len);
         }
 
         /** @return length of current string */
@@ -309,7 +309,7 @@ namespace mongo {
 
         void write( const char* buf, int len) { memcpy( _buf.grow( len ) , buf , len ); }
 
-        void append( const StringData& str ) { memcpy( _buf.grow( str.size() ) , str.data() , str.size() ); }
+        void append( const StringData& str ) { memcpy( _buf.grow( (int)str.size() ) , str.data() , str.size() ); }
 
         StringBuilderImpl& operator<<( const StringData& str ) {
             append( str );
